Move dialogs into dedicated dialogs/ folder
