+++ app/(tabs)/trend.tsx
@@ -1,5 +1,5 @@
 import { Calendar } from '@/components/Calendar';
-import { EditMealDialog } from '@/components/EditMealDialog';
+import { EditMealDialog } from '@/dialogs/EditMealDialog';
 import { MealCard } from '@/components/MealCard';
 import { ProgressCircle } from '@/components/ProgressCircle';
 import { MEAL_META } from '@/constants/meals';
